requirements.txt
streamlit==1.28.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
